DRT-4324 move polling to it's own file, out of Application.
